\name{Airpolution}
\docType{data}
\alias{Airpolution}
\title{Airpolution Data}
\description{
  An airpolution data for testing \code{\link{perars}}.
}
\usage{
  data(Airpolution)
}
\format{
  A time series of 372 observations.
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}
\keyword{datasets}
