\name{Amerikamaru}
\alias{Amerikamaru}
\title{Amerikamaru Data}
\description{
  A multivariate non-stationary test data.
}
\usage{
  data(Amerikamaru)
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}

\keyword{ts}
