\name{bispecData}
\alias{bispecData}
\title{Test Data}
\description{
  A test data for "bispec" and "thirmo".
}
\usage{
  data(bispecData)
}
\source{
  H.Akaike, E.Arahata and T.Ozaki (1976) \emph{Computer Science Monograph, No.6,
   Timsac74 A Time Series Analysis and Control Program Package (2)}.
  The Institute of Statistical Mathematics.
}

\keyword{ts}
