\name{Blsallfood}
\docType{data}
\alias{Blsallfood}
\title{Blsallfood Data}
\description{
  The BLSALLFOOD data. (the Bureau of Labor Statistics, all employees in food industries, January 1967 - December 1979)
}
\usage{
  data(Blsallfood)
}
\format{
  A time series of 156 observations.
}
\source{
  H.Akaike, T.Ozaki, M.Ishiguro, Y.Ogata, G.Kitagawa, Y-H.Tamura, E.Arahata,
  K.Katsura and Y.Tamura (1984) \emph{Computer Science Monographs, Timsac-84
  Part 1.} The Institute of Statistical Mathematics.
}
\keyword{datasets}