\name{Canadianlynx}
\alias{Canadianlynx}
\title{Test Data from canadianlynxes}
\description{
  A test data for "unimar", "unibar", "bsubst" and "exsar".
}
\usage{
  data(Canadianlynx)
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}

\keyword{ts}
