\name{LaborData}
\docType{data}
\alias{LaborData}
\title{Labor force Data}
\description{
  Labor force U.S. unemployed 16 years or over (1972-1978) data.
}
\usage{
  data(LaborData)
}
\format{
  A time series of 72 observations.
}
\source{
  H.Akaike, T.Ozaki, M.Ishiguro, Y.Ogata, G.Kitagawa, Y-H.Tamura, E.Arahata,
  K.Katsura and Y.Tamura (1985)
  \emph{Computer Science Monograph, No.22, Timsac84 Part 1.}
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
