\name{locarData}
\alias{locarData}
\title{Non-stationary Test Data}
\description{
  A non-stationary test data for "mlocar" and "blocar".
}
\usage{
  data(locarData)
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}

\keyword{ts}
