\name{MYE1F}
\docType{data}
\alias{MYE1F}
\title{An earthquake wave data}
\description{
  An earthquake wave data.
}
\usage{
  data(MYE1F)
}
\format{
  A time series of 2600 observations.
}
\source{
  G.Kitagawa (1993) \emph{Time series analysis programing} The Iwanami Computer Science Series.
}

\keyword{datasets}
