\name{nonstData}
\alias{nonstData}
\title{Non-stationary Test Data}
\description{
  A non-stationary test data for "nonst".
}
\usage{
  data(nonstData)
}
\source{
  H.Akaike, E.Arahata and T.Ozaki (1976) \emph{Computer Science Monograph, No.6,
   Timsac74 A Time Series Analysis and Control Program Package (2)}.
  The Institute of Statistical Mathematics.
}

\keyword{ts}
