\name{plot.decomp}
\alias{plot.decomp}
\title{Plot Trend, Seasonal, AR Components and Trading Day Factor}
\description{
  Plot trend component, seasonal component, AR component, noise and trading day
  factor returned by \code{\link{decomp}}.
}
\usage{
\method{plot}{decomp}(x, \dots)
}
\arguments{
  \item{x}{an object of class \code{"decomp"}.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
