\name{Powerplant}
\docType{data}
\alias{Powerplant}
\title{Power Plant Data}
\description{
  A Power plant data for testing \code{\link{mulbar}} and \code{\link{mulmar}}.
}
\usage{
  data(Powerplant)
}
\format{
  A 2-dimensional array with 500 observations on 3 variables.
  \tabular{rl}{
    [,1] \tab command \cr
    [,2] \tab temperature \cr
    [,3] \tab fuel
  }
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}
\keyword{datasets}
