\name{timsac-package}
\alias{timsac-package}
\alias{timsac}
\docType{package}

\title{
Time Series Analysis and Control Program Package
}

\description{
R functions for statistical analysis and control of time series.
}

\details{
This package provides functions for statistical analysis, prediction and control of time series.
For a complete list of functions, use library(help="timsac"). 

 For overview of models and information criteria for model selection, see \url{../doc/timsac-guide_e.pdf}
 or \url{../doc/timsac-guide_j.pdf} (in Japanese).
}

\references{
H.Akaike, E.Arahata and T.Ozaki (1975)
 \emph{Computer Science Monograph, No.5, Timsac74, A Time Series Analysis and Control Program Package (1)}.
 The Institute of Statistical Mathematics.

H.Akaike, E.Arahata and T.Ozaki (1975)
 \emph{Computer Science Monograph, No.6, Timsac74, A Time Series Analysis and Control Program Package (2)}.
 The Institute of Statistical Mathematics.

H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
 \emph{Computer Science Monograph, No.11, Timsac78.} The Institute of Statistical Mathematics.

H.Akaike, T.Ozaki, M.Ishiguro, Y.Ogata, G.Kitagawa, Y-H.Tamura, E.Arahata, K.Katsura and Y.Tamura (1985)
 \emph{Computer Science Monograph, No.22, Timsac84 Part 1}. The Institute of Statistical Mathematics.

H.Akaike and T.Nakagawa (1988)
 \emph{Statistical Analysis and Control of Dynamic Systems.} Kluwer Academic publishers.
}

\keyword{package}


