\name{Airpolution}
\alias{Airpolution}
\title{Airpolution Data}
\description{
  An air polution test data for "perars".
}
\usage{
  data(Airpolution)
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}
\keyword{ts}
