\name{Amerikamaru}
\docType{data}
\alias{Amerikamaru}
\title{Amerikamaru Data}
\description{
  A multivariate non-stationary data for testing \code{\link{blomar}}.
}
\usage{
  data(Amerikamaru)
}
\format{
  A 2-dimensional array with 896 observations on 2 variables.
  \tabular{rl}{
      [, 1] \tab rudder\cr
      [, 2] \tab yawing
  }
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}
\keyword{datasets}
