\name{bispecData}
\docType{data}
\alias{bispecData}
\title{Univariate Test Data}
\description{
  A univariate data for testing \code{\link{bispec}} and \code{\link{thirmo}}.
}
\usage{
  data(bispecData)
}
\format{
  A time series of 1500 observations.
}
\source{
  H.Akaike, E.Arahata and T.Ozaki (1976) \emph{Computer Science Monograph, No.6,
  Timsac74 A Time Series Analysis and Control Program Package (2)}.
 The Institute of Statistical Mathematics.
}
\keyword{datasets}
