\name{Blsallfood}
\alias{Blsallfood}
\title{Blsallfood data}
\description{
  A blsallfood test data for "decomp".
}
\usage{
  data(Blsallfood)
}
\source{
  H.Akaike, T.Ozaki, M.Ishiguro, Y.ogata, G.Kitagawa, Y-H.tamura, E.Arahata, K.Katsura and Y.tamura (1984)
  \emph{Computer Science Monographs, Timsac-84 Part 1.}
  The Institute of Statistical Mathematics.
}

\keyword{ts}
