\name{Canadianlynx}
\docType{data}
\alias{Canadianlynx}
\title{Time series of Canadian lynx data}
\description{
  A time series of Canadian lynx data for testing \code{unimar}, \code{unibar}, \code{bsubst} and \code{exsar}.
}
\usage{
  data(Canadianlynx)
}
\format{
  A time series of 114 observations.
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}

\keyword{datasets}