\name{LaborData}
\alias{LaborData}
\title{Labor force Data}
\description{
  Labor force U.S. unemploved 16 yeaes or over (1972-1978) data.
}
\usage{
  data(LaborData)
}
\source{
  H.Akaike, T.Ozaki, M.Ishiguro, Y.Ogata, G.Kitagawa, Y-H.Tamura, E.Arahata, K.Katsura and Y.Tamura (1985)
  \emph{Computer Science Monograph, No.22, Timsac84 Part 1.}
  The Institute of Statistical Mathematics.
}

\keyword{ts}
