\name{locarData}
\docType{data}
\alias{locarData}
\title{Non-stationary Test Data}
\description{
  A non-stationary data for testing \code{\link{mlocar}} and \code{\link{blocar}}.
}
\usage{
  data(locarData)
}
\format{
  A time series of 1000 observations.
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
