\name{nonstData}
\docType{data}
\alias{nonstData}
\title{Non-stationary Test Data}
\description{
  A non-stationary data for testing \code{\link{nonst}}.
}
\usage{
  data(nonstData)
}
\format{
  A time series of 2100 observations.
}
\source{
  H.Akaike, E.Arahata and T.Ozaki (1976) \emph{Computer Science Monograph, No.6,
   Timsac74 A Time Series Analysis and Control Program Package (2)}.
  The Institute of Statistical Mathematics.
}

\keyword{datasets}
