\name{Powerplant}
\alias{Powerplant}
\title{Power Plant Data}
\description{
  A Power plant test data; 1-ch command, 2-ch temperature, 3-ch fuel.
}
\usage{
  data(Powerplant)
}
\source{
  H.Akaike, G.Kitagawa, E.Arahata and F.Tada (1979)
  \emph{Computer Science Monograph, No.11, Timsac78.}
  The Institute of Statistical Mathematics.
}
\keyword{ts}
